backend/pyproject.toml
[project]
name = "trialpulse-backend"
version = "0.1.0"
description = "TrialPulse — AI-powered clinical trial patient safety and engagement platform"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",

    # Database (SQLModel = SQLAlchemy + Pydantic unified)
    "sqlmodel>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",

    # Redis
    "redis[hiredis]>=5.2.0",

    # LLM Orchestration (LangChain v1 + LangGraph v1 via OpenRouter)
    "langchain>=1.0.0",
    "langchain-core>=0.3.0",
    "langchain-openrouter>=0.1.0",
    "langgraph>=1.0.0",

    # LiveKit + Gemini Live Voice
    "livekit>=0.17.0",
    "livekit-agents>=1.0.0",
    "livekit-plugins-google>=1.0.0",

    # ML / Data
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
    "scipy>=1.14.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # Object storage
    "minio>=7.2.0",

    # Scheduling
    "apscheduler>=3.10.0",

    # Utilities
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py314"
line-length = 100
